JPA 엔티티에 hashCode()와 equals()를 구현해야 할까?
