+++ back-end/src/main/java/org/antop/billiardslove/jpa/entity/CodeId.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.util.Objects;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -13,4 +14,18 @@ import java.io.Serializable;
 public class CodeId implements Serializable {
     private String group;
     private String id;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        CodeId codeId = (CodeId) o;
+        return Objects.equals(group, codeId.group) && Objects.equals(id, codeId.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(group, id);
+    }
+
 }
